move hashing function to atoms/base/hash.h
